+++ triggers/service/BUILD.bazel
@@ -49,7 +49,7 @@ da_scala_library(
         "//libs-scala/ports",
         "//libs-scala/scala-utils",
         "//triggers/runner:trigger-runner-lib",
-        "//triggers/service/auth:oauth-middleware",  # TODO[AH] Extract request/response types
+        "//triggers/service/auth:middleware-api",
         "@maven//:com_chuusai_shapeless_2_12",
         "@maven//:com_github_scopt_scopt_2_12",
         "@maven//:com_lihaoyi_sourcecode_2_12",
@@ -123,8 +123,8 @@ da_scala_library(
         "//libs-scala/resources",
         "//libs-scala/scala-utils",
         "//libs-scala/timer-utils",
-        "//triggers/service/auth:oauth-middleware",
-        "//triggers/service/auth:oauth-test-server",
+        "//triggers/service/auth:oauth2-middleware",
+        "//triggers/service/auth:oauth2-test-server",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_auth0_java_jwt",
         "@maven//:com_typesafe_akka_akka_actor_typed_2_12",
@@ -188,7 +188,7 @@ da_scala_test_suite(
         "//libs-scala/ports",
         "//libs-scala/postgresql-testing",
         "//libs-scala/resources",
-        "//triggers/service/auth:oauth-test-server",
+        "//triggers/service/auth:oauth2-test-server",
         "@maven//:com_typesafe_akka_akka_http_core_2_12",
         "@maven//:eu_rekawek_toxiproxy_toxiproxy_java_2_1_3",
         "@maven//:io_spray_spray_json_2_12",
