+++ triggers/service/auth/src/test/scala/com/daml/auth/middleware/oauth2/TestFixture.scala
@@ -1,13 +1,20 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.oauth.server
+package com.daml.auth.middleware.oauth2
 
 import java.time.{Instant, ZoneId}
+import java.util.Date
 
 import akka.http.scaladsl.Http.ServerBinding
 import akka.http.scaladsl.model.Uri
+import com.auth0.jwt.JWTVerifier.BaseVerification
+import com.auth0.jwt.JWT
+import com.auth0.jwt.algorithms.Algorithm
+import com.auth0.jwt.interfaces.{Clock => Auth0Clock}
 import com.daml.clock.AdjustableClock
+import com.daml.jwt.JwtVerifier
+import com.daml.ledger.api.refinements.ApiTypes
 import com.daml.ledger.api.testing.utils.{
   AkkaBeforeAndAfterAll,
   OwnedResource,
@@ -15,8 +22,8 @@ import com.daml.ledger.api.testing.utils.{
   SuiteResource
 }
 import com.daml.ledger.resources.ResourceContext
+import com.daml.auth.oauth2.test.server.{Config => OAuthConfig}
 import com.daml.ports.Port
-import com.daml.ledger.api.refinements.ApiTypes.Party
 import org.scalatest.Suite
 
 trait TestFixture
@@ -24,7 +31,6 @@ trait TestFixture
     with SuiteResource[(AdjustableClock, ServerBinding, ServerBinding)] {
   self: Suite =>
   protected val ledgerId: String = "test-ledger"
-  protected val applicationId: String = "test-application"
   protected val jwtSecret: String = "secret"
   override protected lazy val suiteResource
     : Resource[(AdjustableClock, ServerBinding, ServerBinding)] = {
@@ -33,21 +39,31 @@ trait TestFixture
       for {
         clock <- Resources.clock(Instant.now(), ZoneId.systemDefault())
         server <- Resources.authServer(
-          Config(
+          OAuthConfig(
             port = Port.Dynamic,
             ledgerId = ledgerId,
             jwtSecret = jwtSecret,
-            parties = Some(Party.subst(Set("Alice", "Bob"))),
-            clock = Some(clock)
+            parties = Some(ApiTypes.Party.subst(Set("Alice", "Bob"))),
+            clock = Some(clock),
           ))
-        client <- Resources.authClient(
-          Client.Config(
+        serverUri = Uri()
+          .withScheme("http")
+          .withAuthority(server.localAddress.getHostString, server.localAddress.getPort)
+        client <- Resources.authMiddleware(
+          Config(
             port = Port.Dynamic,
-            authServerUrl = Uri()
-              .withScheme("http")
-              .withAuthority(server.localAddress.getHostString, server.localAddress.getPort),
-            clientId = "test-client",
-            clientSecret = "test-client-secret"
+            oauthAuth = serverUri.withPath(Uri.Path./("authorize")),
+            oauthToken = serverUri.withPath(Uri.Path./("token")),
+            clientId = "middleware",
+            clientSecret = "middleware-secret",
+            tokenVerifier = new JwtVerifier(
+              JWT
+                .require(Algorithm.HMAC256(jwtSecret))
+                .asInstanceOf[BaseVerification]
+                .build(new Auth0Clock {
+                  override def getToday: Date = Date.from(clock.instant())
+                })
+            )
           ))
       } yield { (clock, server, client) }
     )
