+++ triggers/service/auth/src/test/scala/com/daml/auth/oauth2/test/server/Resources.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.oauth.server
+package com.daml.auth.oauth2.test.server
 
 import java.time.{Clock, Duration, Instant, ZoneId}
